rename AudioFile/{AudioFileFormat => format}.hpp
